requirements-api.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
grpcio>=1.60.0
httpx>=0.27.0
